MVP to MVVM
