--- app/src/main/java/pro/fateeva/chuchasdictionarymvp/interactor/Interactor.kt
@@ -1,7 +1,0 @@
-package pro.fateeva.chuchasdictionarymvp.interactor
-
-import io.reactivex.Observable
-
-interface Interactor<T> {
-    fun getData(word: String): Observable<T>
-}
