--- app/src/main/java/pro/fateeva/chuchasdictionarymvp/presenter/Presenter.kt
@@ -1,10 +1,0 @@
-package pro.fateeva.chuchasdictionarymvp.presenter
-
-import pro.fateeva.chuchasdictionarymvp.model.AppState
-import pro.fateeva.chuchasdictionarymvp.view.FragmentView
-
-interface Presenter<A: AppState, V: FragmentView> {
-    fun getData(word: String)
-    fun attachFragment(view: V)
-    fun detachFragment(view: V)
-}
