+++ app/src/main/java/pro/fateeva/chuchasdictionarymvp/usecase/ListOfWordsUseCase.kt
@@ -1,16 +1,14 @@
-package pro.fateeva.chuchasdictionarymvp.interactor
+package pro.fateeva.chuchasdictionarymvp.usecase
 
 import io.reactivex.Observable
 import pro.fateeva.chuchasdictionarymvp.model.AppState
 import pro.fateeva.chuchasdictionarymvp.model.Word
 import pro.fateeva.chuchasdictionarymvp.repository.Repository
 
-class InteractorImpl(
+class ListOfWordsUseCase(
     private val remoteRepository: Repository<List<Word>>
-) : Interactor<AppState> {
-
-    override fun getData(word: String): Observable<AppState> {
-       return remoteRepository.getData(word).map { AppState.Success(it) }
+) {
+    fun getData(word: String): Observable<AppState> {
+        return remoteRepository.getData(word).map { AppState.Success(it) }
     }
-
 }
