--- app/src/main/java/pro/fateeva/chuchasdictionarymvp/view/FragmentView.kt
@@ -1,7 +1,0 @@
-package pro.fateeva.chuchasdictionarymvp.view
-
-import pro.fateeva.chuchasdictionarymvp.model.AppState
-
-interface FragmentView  {
-    fun renderData(appState: AppState)
-}
